pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[project]
name = "service" # TODO: service name
version = "0.0.1"
description = ""
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "fastapi[all]>=0.115.8",
  "fastapi-pagination>=0.12.34",
  "google-cloud-logging>=3.11.4",
  "greenlet>=3.1.1",
  "opentelemetry-api>=1.30",
  "opentelemetry-exporter-gcp-logging>=1.9.0a0",
  "opentelemetry-exporter-gcp-trace>=1.9",
  "opentelemetry-propagator-gcp>=1.9",
  "opentelemetry-sdk>=1.30",
  "psycopg[binary]>=3.2.9",
  "pydantic-settings>=2.7.1",
  "sqlmodel>=0.0.22",
  "wrapt>=2.0.1",
]

scripts.app = "src:server"

[dependency-groups]
dev = [
  "alembic>=1.16.4",
  "coverage>=7.6.10",
  "pre-commit>=4.1",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6",
  "ruff>=0.9.4",
]

[tool.hatch.build.targets.wheel]
packages = [ "src" ]

[tool.uv]
package = true

[tool.ruff]
line-length = 80

lint.extend-select = [ "ASYNC", "F", "FAST", "I", "RUF", "UP" ]
